Added UI for verification management, assignments and disputes
