+++ src/app/(pages)/(dashboard)/booking-management/booking-disputes/page.tsx
@@ -1,7 +1,9 @@
+import BookingsTable from "@/src/components/booking-mgt/tables/bookings";
+
 export default function BookingDisputesPage() {
     return (
         <h2 className="">
-            BookingDisputes
+            <BookingsTable />
         </h2>
     );
 }
