+++ src/app/(pages)/(dashboard)/property-management/create/page.tsx
@@ -1,4 +1,4 @@
-import CreatePropertyView from "@/src/components/properties-mgt/CreatePropertyView";
+import CreatePropertyView from "@/src/components/properties-mgt/all-properties/CreatePropertyView";
 
 export default function CreateProperty(){
     return(
