+++ src/app/(pages)/(dashboard)/property-management/manage-verifications/[verificationId]/page.tsx
@@ -1,15 +1,15 @@
 "use client";
 
+import VerifcationDetails from "@/src/components/properties-mgt/manage-verifications/VerificationDetails";
 import { useParams } from "next/navigation";
-import PropertyDetailsView from "@/src/components/properties-mgt/all-properties/PropertyDetailsView";
 
 export default function PropertyDetail() {
     const params = useParams();
     
     return (
         <div className="w-full">
-            <PropertyDetailsView
-                propertyId={Number(params?.propertyId)}
+            <VerifcationDetails
+                verificationId={Number(params?.verificationId)}
             />
         </div>
     );
