+++ src/app/(pages)/(dashboard)/property-management/manage-verifications/page.tsx
@@ -1,8 +1,10 @@
+import VerificationsTable from "@/src/components/properties-mgt/tables/verifications";
+
 export default function ManageVerificationPage() {
     return (
-        <h2 className="">
-            Manage verification
-        </h2>
+        <div className="w-full">
+            <VerificationsTable />
+        </div>
     );
 }
