+++ src/components/booking-mgt/tables/bookings.tsx
@@ -1,23 +1,24 @@
 'use client'
 
+import { formatDate, formatMoney } from "@/src/lib/utils";
+import { DotsIcon, FilterIcon, PrinterIcon, SearchIcon } from "../../icons";
+import { GetAllBookings } from "@/src/lib/request-handlers/bookingMgt";
 import { useEffect, useRef, useState } from "react";
-import { ArrowIcon, DotsIcon, FilterIcon, PrinterIcon, SearchIcon } from "../icons";
-import { useRouter } from "next/navigation";
-import { IProperty, PropertyType } from "./types";
-import { GetAllProperties } from "@/src/lib/request-handlers/propertyMgt";
-import Loader from "../loader";
-import { PAGE_ROUTES } from "@/src/lib/routes/page_routes";
-import { BookingBadge } from "../badge";
-import { formatDate } from "@/src/lib/utils";
-import TablePagination from "../TablePagination";
+import { IBooking } from "../types";
+import { BookingBadge } from "../../badge";
+import TablePagination from "../../TablePagination";
+import Loader from "@/src/components/loader";
 import { LuEye } from "react-icons/lu";
 import { HiOutlinePencilAlt } from "react-icons/hi";
+import { useRouter } from "next/navigation";
+import { PAGE_ROUTES } from "@/src/lib/routes/page_routes";
 
-export default function ListPropertiesView() {
+export default function BookingsTable() {
+
     const [page, setPage] = useState<number>(1);
     const [searchTerm , setSearchTerm] = useState<string>("");
-    const { data: properties, isLoading } = GetAllProperties(page, 10, searchTerm);
-    const [propertyList, setPropertyList] = useState<IProperty[]>(properties?.data?.data?.data);
+    const { data: bookings, isLoading } = GetAllBookings(page, 10, searchTerm);
+    const [bookingList, setBookingList] = useState<IBooking[]>(bookings?.data?.data?.data);
     const router = useRouter();
 
     const [selectedRow, setSelectedRow] = useState<number>(0);
@@ -29,17 +30,17 @@ export default function ListPropertiesView() {
             label: "View",
             Icon: <LuEye />,
             onClick: () => router.push(
-                PAGE_ROUTES.dashboard.propertyManagement.allProperties.details(propertyList[selectedRow].id)
+                PAGE_ROUTES.dashboard.bookingManagement.bookings.details(bookingList[selectedRow].id)
             ),
         },
         {
             label: "Edit",
             Icon: <HiOutlinePencilAlt />,
             onClick: () => router.push(
-                `${PAGE_ROUTES.dashboard.propertyManagement.allProperties.details(propertyList[selectedRow].id)}?edit=true`
+                `${PAGE_ROUTES.dashboard.bookingManagement.bookings.details(bookingList[selectedRow].id)}?edit=true`
             ),
         }
-    ]
+    ];
 
     // Handle click outside modal
     useEffect(() => {
@@ -60,20 +61,23 @@ export default function ListPropertiesView() {
         setModalPosition({ top: rect.bottom + window.scrollY, left: rect.left + window.scrollX });
     };
 
+
     useEffect(() => {
-        setPropertyList(properties?.data?.data?.data);
-    }, [properties])
+        setBookingList(bookings?.data?.data?.data);
+    }, [bookings])
+
 
     return (
+ 
         <div className="w-full p-10">
-            <div className="w-full border border-zinc-500/20 bg-white rounded-xl px-6 py-7 min-h-[70vh] flex flex-col items-center">
-                
+            <div className="w-full border border-zinc-500/20 bg-white rounded-xl px-6 py-7 min-h-[72vh] flex flex-col items-center">
+
                 
                 <div className="w-full flex justify-between items-center">
                     <div className="w-[80%] flex items-center gap-5">
-                        <p className="text-2xl font-medium mr-10">Property Listings</p>
+                        <p className="text-2xl font-medium mr-10">All Bookings</p>
                         <div className="relative w-[40%]">
-                            <input type="text" value={searchTerm} onChange={(e) => setSearchTerm(e.target.value)} className="border border-zinc-500/20 bg-background rounded-lg w-full h-10 p-3 pl-10" placeholder="Search properties "/>
+                            <input type="text" value={searchTerm} onChange={(e) => setSearchTerm(e.target.value)} className="border border-zinc-500/20 bg-background rounded-lg w-full h-10 p-3 pl-10" placeholder="Search bookings "/>
                             <SearchIcon className="absolute top-[25%] left-3 w-5" color="black" />
                         </div>
                         <div className="flex justify-center items-center gap-2 border border-zinc-500/20 bg-background hover:bg-zinc-300/70 cursor-pointer rounded-lg h-10 px-3 w-[10%]">
@@ -91,6 +95,7 @@ export default function ListPropertiesView() {
                     </button>
                 </div>
 
+
                 {
                     isLoading ?
                     <Loader />
@@ -98,9 +103,9 @@ export default function ListPropertiesView() {
                     <div className="w-full mt-6">
                         <table className="w-full border-collapse">
                             <thead className="">
-                                <tr className="text-teal-600 text-[12px]">
+                                <tr className="text-teal-600 text-[13px]">
                                     <th className="bg-[#0280901A] h-10 p-5 flex justify-start items-center gap-3 rounded-tl-xl rounded-bl-xl font-medium w-full">
-                                        {/* <input 
+                                        {/* <input
                                             type="checkbox"
                                             className={`
                                                 size-4 border-2 border-teal-600 rounded-md bg-transparent appearance-none
@@ -108,12 +113,12 @@ export default function ListPropertiesView() {
                                             `}
                                         /> */}
                                         <p>
-                                            Property ID
+                                            Booking ID
                                         </p>
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-left">
                                         <p>
-                                            Property Name
+                                            Property
                                         </p>
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-left">
@@ -123,33 +128,35 @@ export default function ListPropertiesView() {
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-left">
                                         <p>
-                                            {"Owner's Name"}
-                                        </p>
-                                    </th>
-                                    <th className="bg-[#0280901A] h-10 font-medium text-center">
-                                        <p>
-                                            Verification status
+                                            {"Guest's Name"}
                                         </p>
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-left">
                                         <p>
-                                            Assigned agent
+                                            Price
                                         </p>
                                     </th>
-                                    <th className="bg-[#0280901A] h-10 font-medium  text-left">
+                                    <th className="bg-[#0280901A] h-10 font-medium text-center">
+                                        <p>
+                                            Status
+                                        </p>
+                                    </th>
+                                    <th className="bg-[#0280901A] h-10 font-medium  text-center">
                                         <p className="pr-2">
-                                            Submission date
+                                            Booking date
                                         </p>
                                     </th>
-                                    <th className="bg-[#0280901A] h-10 rounded-tr-xl rounded-br-xl">{' '}</th>
+                                    <th className="bg-[#0280901A] h-10 ">{' '}</th>
                                     <th className="bg-[#0280901A] h-10 rounded-tr-xl rounded-br-xl">{' '}</th>
                                 </tr>
                             </thead>
                             <tbody className="text-[13px]">
                                 {
-                                    propertyList &&
-                                    propertyList?.map((property, index) => (
-                                        <tr key={index} className="hover:bg-background/50 cursor-pointer"  onClick={() => router.push(PAGE_ROUTES.dashboard.propertyManagement.allProperties.details(property?.id))}>
+                                    bookingList &&
+                                    bookingList?.map((booking, index) => (
+                                        <tr key={index} className="hover:bg-background/50 cursor-pointer" 
+                                            onClick={() => router.push(PAGE_ROUTES.dashboard.bookingManagement.bookings.details(booking?.id))} 
+                                        >
                                             <td className="flex items-center px-5 py-4 gap-3 border-b-2 border-b-gray-200">
                                                 {/* <input 
                                                     type="checkbox"
@@ -158,42 +165,32 @@ export default function ListPropertiesView() {
                                                         checked:bg-zinc-800 checked:border-zinc-800 checked:text-zinc-200
                                                     `}
                                                 /> */}
-                                                <p className="pt-1">
-                                                    APT25-{property?.id}
+                                                <p className="pt-1 truncate max-w-36">
+                                                    {booking?.bookingId}
                                                 </p>  
                                             </td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <p className="pt-1 truncate max-w-[13rem]">
-                                                    {property?.name}
+                                                    {booking?.unit?.name}
                                                 </p>   
                                             </td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <p className="pt-1 font-medium">
-                                                    {property?.propertyType}
+                                                    {booking?.unit?.property?.propertyType}
                                                 </p>
                                             </td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <p className="pt-1 ">
-                                                    {`${property?.owner?.profile?.firstName??'John'} ${property?.owner?.profile?.lastName??'Oderinde'}`}
+                                                    {`${booking?.user?.profile?.firstName??'Fola'} ${booking?.user?.profile?.lastName??'Ogunleye'}`}
                                                 </p>
                                             </td>
+                                            <td className="border-b-2 border-b-gray-200">{`₦ ${formatMoney(Number(booking?.totalPrice))}`}</td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <div className="py-2.5 w-2/3 m-auto text-center">
-                                                    <BookingBadge 
-                                                        status={property?.isVerified ? 'Verified' : 'Unverified'} 
-                                                        textColour={property?.isVerified ? "text-[#028090]" : "text-red-600" }
-                                                        backgroundColour={property?.isVerified ? "bg-[#0280901A]" : "bg-red-200"}
-                                                    />
+                                                    <BookingBadge status={booking?.status} classNames="" />
                                                 </div>
                                             </td>
-                                            <td className="border-b-2 border-b-gray-200">
-                                                <p className="pt-1 ">
-                                                    {`${property?.agent?.profile?.firstName??'Thomas'} ${property?.agent?.profile?.lastName??'Alexander'}`}
-                                                </p>
-                                            </td>
-                                            <td className="border-b-2 border-b-gray-200">
-                                                {formatDate(property?.verifications?.createdAt??"02-03-2024")}
-                                            </td>
+                                            <td className="border-b-2 border-b-gray-200 text-center">{formatDate(booking.startDate)}</td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <div className="flex justify-center items-center w-fit" onClick={(event) => handleDotsClick(event, index)}>
                                                     <DotsIcon className="w-5 ml-12 cursor-pointer " color="gray" />
@@ -206,8 +203,8 @@ export default function ListPropertiesView() {
                         </table>
                     </div>
                 }
-            </div>
 
+            </div>
 
             {/* Modal */}
             {selectedRow !== null && modalPosition && (
@@ -228,18 +225,19 @@ export default function ListPropertiesView() {
                     ))}
                 </div>
             )}
-            
+
             {
-                !isLoading && propertyList &&
+                !isLoading && bookingList &&
                 <TablePagination
-                    total={properties?.data?.data?.meta?.total}
+                    total={bookings?.data?.data?.meta?.total}
                     currentPage={page}
                     setPage={setPage}
-                    firstPage={properties?.data?.data?.meta?.firstPage}
+                    firstPage={bookings?.data?.data?.meta?.firstPage}
                     itemsPerPage={10}
                 />
-            }   
-
+            }          
+            
         </div>
+            
     );
 };
