+++ src/components/properties-mgt/all-properties/CreatePropertyView.tsx
@@ -4,21 +4,21 @@ import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 import { FaRegBuilding } from "react-icons/fa";
 import { FaMapLocationDot } from "react-icons/fa6";
 import { SlLocationPin } from "react-icons/sl";
-import CustomDropdown from "../ui/customDropdown";
-import { IAmenity, ICreateProperty, MediaType, PropertyType } from "./types";
-import CustomFilterDropdown from "../ui/customFilterDropDown";
-import CustomCheckbox from "../ui/customCheckbox";
-import MultipleChoice from "../ui/MultipleChoice";
+import CustomDropdown from "../../ui/customDropdown";
+import { IAmenity, ICreateProperty, MediaType, PropertyType } from "../types";
+import CustomFilterDropdown from "../../ui/customFilterDropDown";
+import CustomCheckbox from "../../ui/customCheckbox";
+import MultipleChoice from "../../ui/MultipleChoice";
 import { ALL_COUNTRIES } from "@/src/data/countries";
 import { FaPlus } from "react-icons/fa6";
-import CustomDropzone from "../ui/CustomDropzone";
+import CustomDropzone from "../../ui/CustomDropzone";
 import { useFormik } from 'formik';
 import { CreateAmenity, CreateProperty, UploadPropertyMedia } from "@/src/lib/request-handlers/propertyMgt";
 import { useAuth } from "@/src/hooks/useAuth";
-import Spinner from "../ui/Spinner";
+import Spinner from "../../ui/Spinner";
 import { GetAmenities } from "@/src/lib/request-handlers/propertyMgt";
 import { fixedAmenities } from "@/src/data/amenities";
-import CustomModal from "../ui/CustomModal";
+import CustomModal from "../../ui/CustomModal";
 import { UserRole } from "@/src/lib/enums";
 import { useRouter } from "next/navigation";
 import { PAGE_ROUTES } from "@/src/lib/routes/page_routes";
