+++ src/components/properties-mgt/all-properties/EditPropertyView.tsx
@@ -3,27 +3,27 @@
 import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 import { FaRegBuilding } from "react-icons/fa";
 import { FaMapLocationDot } from "react-icons/fa6";
-import { TrashIcon } from "../icons";
+import { TrashIcon } from "../../icons";
 import { SlLocationPin } from "react-icons/sl";
-import CustomDropdown from "../ui/customDropdown";
-import { IAmenity, IProperty, IPropertyMedia, IUpdateProperty, MediaType, PropertyType } from "./types";
-import CustomFilterDropdown from "../ui/customFilterDropDown";
-import CustomCheckbox from "../ui/customCheckbox";
-import MultipleChoice from "../ui/MultipleChoice";
+import CustomDropdown from "../../ui/customDropdown";
+import { IAmenity, IProperty, IPropertyMedia, IUpdateProperty, MediaType, PropertyType } from "../types";
+import CustomFilterDropdown from "../../ui/customFilterDropDown";
+import CustomCheckbox from "../../ui/customCheckbox";
+import MultipleChoice from "../../ui/MultipleChoice";
 import { ALL_COUNTRIES } from "@/src/data/countries";
 import { FaPlus } from "react-icons/fa6";
 import { IoCloudUploadOutline } from "react-icons/io5";
 import Image from "next/image";
 import { showAlert } from "@/src/lib/slices/alertDialogSlice";
 import { useDispatch } from "react-redux";
-import CustomDropzone from "../ui/CustomDropzone";
+import CustomDropzone from "../../ui/CustomDropzone";
 import { useFormik } from 'formik';
 import { DeleteProperty, FeatureProperty, UpdateProperty, UploadPropertyMedia } from "@/src/lib/request-handlers/propertyMgt";
 import { useAuth } from "@/src/hooks/useAuth";
 import { UserRole } from "@/src/lib/enums";
-import Spinner from "../ui/Spinner";
+import Spinner from "../../ui/Spinner";
 import { CreateAmenityForm } from "./CreatePropertyView";
-import CustomModal from "../ui/CustomModal";
+import CustomModal from "../../ui/CustomModal";
 import { useRouter, useSearchParams } from "next/navigation";
 import { PAGE_ROUTES } from "@/src/lib/routes/page_routes";
 import toast from "react-hot-toast";
