+++ src/components/properties-mgt/tables/properties.tsx
@@ -1,24 +1,23 @@
 'use client'
 
-import { formatDate, formatMoney } from "@/src/lib/utils";
-import { DotsIcon, FilterIcon, PrinterIcon, SearchIcon } from "../icons";
-import { GetAllBookings } from "@/src/lib/request-handlers/bookingMgt";
 import { useEffect, useRef, useState } from "react";
-import { IBooking } from "./types";
-import { BookingBadge } from "../badge";
-import TablePagination from "../TablePagination";
-import Loader from "@/src/components/loader";
+import { ArrowIcon, DotsIcon, FilterIcon, PrinterIcon, SearchIcon } from "../../icons";
+import { useRouter } from "next/navigation";
+import { IProperty, PropertyType } from "../types";
+import { GetAllProperties } from "@/src/lib/request-handlers/propertyMgt";
+import Loader from "../../loader";
+import { PAGE_ROUTES } from "@/src/lib/routes/page_routes";
+import { BookingBadge } from "../../badge";
+import { formatDate } from "@/src/lib/utils";
+import TablePagination from "../../TablePagination";
 import { LuEye } from "react-icons/lu";
 import { HiOutlinePencilAlt } from "react-icons/hi";
-import { useRouter } from "next/navigation";
-import { PAGE_ROUTES } from "@/src/lib/routes/page_routes";
 
-export default function ListBookingView() {
-
+export default function PropertiesTable() {
     const [page, setPage] = useState<number>(1);
     const [searchTerm , setSearchTerm] = useState<string>("");
-    const { data: bookings, isLoading } = GetAllBookings(page, 10, searchTerm);
-    const [bookingList, setBookingList] = useState<IBooking[]>(bookings?.data?.data?.data);
+    const { data: properties, isLoading } = GetAllProperties(page, 10, searchTerm);
+    const [propertyList, setPropertyList] = useState<IProperty[]>(properties?.data?.data?.data);
     const router = useRouter();
 
     const [selectedRow, setSelectedRow] = useState<number>(0);
@@ -29,18 +28,24 @@ export default function ListBookingView() {
         {
             label: "View",
             Icon: <LuEye />,
-            onClick: () => router.push(
-                PAGE_ROUTES.dashboard.bookingManagement.bookings.details(bookingList[selectedRow].id)
-            ),
+            onClick: () => {
+                router.push(
+                    PAGE_ROUTES.dashboard.propertyManagement.allProperties.details(propertyList[selectedRow].id)
+                )
+                setSelectedRow(0)
+            },
         },
         {
             label: "Edit",
             Icon: <HiOutlinePencilAlt />,
-            onClick: () => router.push(
-                `${PAGE_ROUTES.dashboard.bookingManagement.bookings.details(bookingList[selectedRow].id)}?edit=true`
-            ),
+            onClick: () => {
+                router.push(
+                    `${PAGE_ROUTES.dashboard.propertyManagement.allProperties.details(propertyList[selectedRow].id)}?edit=true`
+                )
+                setSelectedRow(0)
+            },
         }
-    ];
+    ]
 
     // Handle click outside modal
     useEffect(() => {
@@ -61,23 +66,20 @@ export default function ListBookingView() {
         setModalPosition({ top: rect.bottom + window.scrollY, left: rect.left + window.scrollX });
     };
 
-
     useEffect(() => {
-        setBookingList(bookings?.data?.data?.data);
-    }, [bookings])
-
+        setPropertyList(properties?.data?.data?.data);
+    }, [properties])
 
     return (
- 
         <div className="w-full p-10">
-            <div className="w-full border border-zinc-500/20 bg-white rounded-xl px-6 py-7 min-h-[72vh] flex flex-col items-center">
-
+            <div className="w-full border border-zinc-500/20 bg-white rounded-xl px-6 py-7 min-h-[70vh] flex flex-col items-center">
+                
                 
                 <div className="w-full flex justify-between items-center">
                     <div className="w-[80%] flex items-center gap-5">
-                        <p className="text-2xl font-medium mr-10">All Bookings</p>
+                        <p className="text-2xl font-medium mr-10">Property Listings</p>
                         <div className="relative w-[40%]">
-                            <input type="text" value={searchTerm} onChange={(e) => setSearchTerm(e.target.value)} className="border border-zinc-500/20 bg-background rounded-lg w-full h-10 p-3 pl-10" placeholder="Search bookings "/>
+                            <input type="text" value={searchTerm} onChange={(e) => setSearchTerm(e.target.value)} className="border border-zinc-500/20 bg-background rounded-lg w-full h-10 p-3 pl-10" placeholder="Search properties "/>
                             <SearchIcon className="absolute top-[25%] left-3 w-5" color="black" />
                         </div>
                         <div className="flex justify-center items-center gap-2 border border-zinc-500/20 bg-background hover:bg-zinc-300/70 cursor-pointer rounded-lg h-10 px-3 w-[10%]">
@@ -95,7 +97,6 @@ export default function ListBookingView() {
                     </button>
                 </div>
 
-
                 {
                     isLoading ?
                     <Loader />
@@ -103,9 +104,9 @@ export default function ListBookingView() {
                     <div className="w-full mt-6">
                         <table className="w-full border-collapse">
                             <thead className="">
-                                <tr className="text-teal-600 text-[13px]">
+                                <tr className="text-teal-600 text-[12px]">
                                     <th className="bg-[#0280901A] h-10 p-5 flex justify-start items-center gap-3 rounded-tl-xl rounded-bl-xl font-medium w-full">
-                                        {/* <input
+                                        {/* <input 
                                             type="checkbox"
                                             className={`
                                                 size-4 border-2 border-teal-600 rounded-md bg-transparent appearance-none
@@ -113,12 +114,12 @@ export default function ListBookingView() {
                                             `}
                                         /> */}
                                         <p>
-                                            Booking ID
+                                            Property ID
                                         </p>
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-left">
                                         <p>
-                                            Property
+                                            Property Name
                                         </p>
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-left">
@@ -128,35 +129,33 @@ export default function ListBookingView() {
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-left">
                                         <p>
-                                            {"Guest's Name"}
-                                        </p>
-                                    </th>
-                                    <th className="bg-[#0280901A] h-10 font-medium text-left">
-                                        <p>
-                                            Price
+                                            {"Owner's Name"}
                                         </p>
                                     </th>
                                     <th className="bg-[#0280901A] h-10 font-medium text-center">
                                         <p>
-                                            Status
+                                            Verification status
                                         </p>
                                     </th>
-                                    <th className="bg-[#0280901A] h-10 font-medium  text-center">
+                                    <th className="bg-[#0280901A] h-10 font-medium text-left">
+                                        <p>
+                                            Assigned agent
+                                        </p>
+                                    </th>
+                                    <th className="bg-[#0280901A] h-10 font-medium  text-left">
                                         <p className="pr-2">
-                                            Booking date
+                                            Submission date
                                         </p>
                                     </th>
-                                    <th className="bg-[#0280901A] h-10 ">{' '}</th>
+                                    <th className="bg-[#0280901A] h-10 rounded-tr-xl rounded-br-xl">{' '}</th>
                                     <th className="bg-[#0280901A] h-10 rounded-tr-xl rounded-br-xl">{' '}</th>
                                 </tr>
                             </thead>
                             <tbody className="text-[13px]">
                                 {
-                                    bookingList &&
-                                    bookingList?.map((booking, index) => (
-                                        <tr key={index} className="hover:bg-background/50 cursor-pointer" 
-                                            onClick={() => router.push(PAGE_ROUTES.dashboard.bookingManagement.bookings.details(booking?.id))} 
-                                        >
+                                    propertyList &&
+                                    propertyList?.map((property, index) => (
+                                        <tr key={index} className="hover:bg-background/50 cursor-pointer"  onClick={() => router.push(PAGE_ROUTES.dashboard.propertyManagement.allProperties.details(property?.id))}>
                                             <td className="flex items-center px-5 py-4 gap-3 border-b-2 border-b-gray-200">
                                                 {/* <input 
                                                     type="checkbox"
@@ -165,32 +164,42 @@ export default function ListBookingView() {
                                                         checked:bg-zinc-800 checked:border-zinc-800 checked:text-zinc-200
                                                     `}
                                                 /> */}
-                                                <p className="pt-1 truncate max-w-36">
-                                                    {booking?.bookingId}
+                                                <p className="pt-1">
+                                                    APT25-{property?.id}
                                                 </p>  
                                             </td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <p className="pt-1 truncate max-w-[13rem]">
-                                                    {booking?.unit?.name}
+                                                    {property?.name}
                                                 </p>   
                                             </td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <p className="pt-1 font-medium">
-                                                    {booking?.unit?.property?.propertyType}
+                                                    {property?.propertyType}
                                                 </p>
                                             </td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <p className="pt-1 ">
-                                                    {`${booking?.user?.profile?.firstName??'Fola'} ${booking?.user?.profile?.lastName??'Ogunleye'}`}
+                                                    {`${property?.owner?.profile?.firstName??'John'} ${property?.owner?.profile?.lastName??'Oderinde'}`}
                                                 </p>
                                             </td>
-                                            <td className="border-b-2 border-b-gray-200">{`₦ ${formatMoney(Number(booking?.totalPrice))}`}</td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <div className="py-2.5 w-2/3 m-auto text-center">
-                                                    <BookingBadge status={booking?.status} classNames="" />
+                                                    <BookingBadge 
+                                                        status={property?.isVerified ? 'Verified' : 'Unverified'} 
+                                                        textColour={property?.isVerified ? "text-[#028090]" : "text-red-600" }
+                                                        backgroundColour={property?.isVerified ? "bg-[#0280901A]" : "bg-red-200"}
+                                                    />
                                                 </div>
                                             </td>
-                                            <td className="border-b-2 border-b-gray-200 text-center">{formatDate(booking.startDate)}</td>
+                                            <td className="border-b-2 border-b-gray-200">
+                                                <p className="pt-1 ">
+                                                    {`${property?.agent?.profile?.firstName??'Thomas'} ${property?.agent?.profile?.lastName??'Alexander'}`}
+                                                </p>
+                                            </td>
+                                            <td className="border-b-2 border-b-gray-200">
+                                                {formatDate(property?.verifications?.createdAt??"02-03-2024")}
+                                            </td>
                                             <td className="border-b-2 border-b-gray-200">
                                                 <div className="flex justify-center items-center w-fit" onClick={(event) => handleDotsClick(event, index)}>
                                                     <DotsIcon className="w-5 ml-12 cursor-pointer " color="gray" />
@@ -203,11 +212,11 @@ export default function ListBookingView() {
                         </table>
                     </div>
                 }
-
             </div>
 
+
             {/* Modal */}
-            {selectedRow !== null && modalPosition && (
+            {selectedRow !== 0 && modalPosition && (
                 <div
                     ref={modalRef}
                     className="absolute bg-white shadow-md rounded-md z-50 border border-gray-300 w-[9em]"
@@ -225,19 +234,18 @@ export default function ListBookingView() {
                     ))}
                 </div>
             )}
-
+            
             {
-                !isLoading && bookingList &&
+                !isLoading && propertyList &&
                 <TablePagination
-                    total={bookings?.data?.data?.meta?.total}
+                    total={properties?.data?.data?.meta?.total}
                     currentPage={page}
                     setPage={setPage}
-                    firstPage={bookings?.data?.data?.meta?.firstPage}
+                    firstPage={properties?.data?.data?.meta?.firstPage}
                     itemsPerPage={10}
                 />
-            }          
-            
+            }   
+
         </div>
-            
     );
 };
